refactor(core): one module per resource under store/
`skein-store.ts` had grown past 1,100 lines around six independent per-resource
sections, and adding a seventh would have made it worse rather than better. Each
section moves to its own module — assistants, threads, runs, store items, crons,
idempotency — plus `pagination.ts` and `metadata-match.ts` for the two pieces
every resource shares.

`skein-store.ts` stays the composition point: it declares `SkeinStore` and
`SkeinStoreSnapshot` and re-exports every part, so `src/index.ts` and any deep
importer keep working. Pure move — no contract changes, no behaviour changes, and
the exported surface is unchanged.

Records the practice in docs/code-practices.md so the next large file gets split
before it grows rather than after.
